refactor(coding-agent): centralize thinking levels
